+++ gamemodes/terrortown/entities/weapons/weapon_tttbase.lua
@@ -697,19 +697,14 @@ if CLIENT then
             secondary = secondary and ParT(secondary, translate_params)
         end
 
-        --find mouse keys in the texts to add respective icons
-        primary_key = primary and string.find(primary, "MOUSE1") and Key("+attack", "MOUSE1") or nil
-        secondary_key = secondary and string.find(secondary, "MOUSE2") and Key("+attack2", "MOUSE2")
-            or nil
-
         self:ClearHUDHelp()
 
         if primary then
-            self:AddHUDHelpLine(primary, primary_key)
+            self:AddHUDHelpLine(primary, Key("+attack", "MOUSE1"))
         end
 
         if secondary then
-            self:AddHUDHelpLine(secondary, secondary_key)
+            self:AddHUDHelpLine(secondary, Key("+attack2", "MOUSE2"))
         end
     end
 
@@ -871,9 +866,10 @@ if CLIENT then
         weaponrenderer.RenderWorldModel(self, self, self.customWorldModelElements, self:GetOwner())
     end
 
+    local weaponIsHidden = false
+
     ---
-    -- Allows you to modify viewmodel while the weapon in use before it is drawn.
-    -- @warning This hook only works if you haven't overridden @{GM:PreDrawViewModel}.
+    -- Allows you to modify the viewmodel of the weapon in use before it is drawn.
     -- @param Entity viewModel This is the view model entity before it is drawn
     -- @param Player ply The the owner of the view model
     -- @param Weapon wep This is the weapon that is from the view model
@@ -893,17 +889,38 @@ if CLIENT then
         if wep.UseHands and wep.ShowDefaultViewModel == false then
             viewModel:SetMaterial("vgui/hsv")
 
+            -- trigger a texture reset after the view model is drawn
+            weaponIsHidden = true
+
             return
         end
-
-        -- default case: Normal view model texture is used and view model draw is defined
-        -- with the SWEP.ShowDefaultViewModel variable
-        viewModel:SetMaterial("")
 
         -- only return something if we actually want to hide it because otherwise the SWEP
         -- hook is never called even if the view model is rendered
         if wep.ShowDefaultViewModel == false then
             return true
+        end
+    end)
+
+    ---
+    -- Allows you to modify the viewmodel of the weapon in use after it is drawn.
+    -- @param Entity viewModel This is the view model entity before it is drawn
+    -- @param Player ply The the owner of the view model
+    -- @param Weapon wep This is the weapon that is from the view model
+    -- @realm client
+    hook.Add("PostDrawViewModel", "TTT2ViewModelHiderReset", function(viewModel, ply, wep)
+        -- default case: Normal view model texture is used and view model draw is defined
+        -- with the SWEP.ShowDefaultViewModel variable
+
+        -- note: we only reset the material to the default material if it was previously set to
+        -- the invisible debug material. That way it is only applied to view models that are
+        -- intended to be invisible where it doesn't matter if it messes up their materials.
+        -- This is done because some weapons set custom materials to the view model (e.g. the
+        -- zombie perk bottles) and always resetting it makes the texture the error texture.
+        if weaponIsHidden then
+            viewModel:SetMaterial("")
+
+            weaponIsHidden = false
         end
     end)